Logging changes for package reporting and Data Model for IX
Changes in this pull request: - RR Logging IX specific fields (as per https://github.corp.inmobi.com/channel-adserve/cas-thrift/pull/84) - IXPackage id is defined as Integer in the tables. Using Integer instead of long.  - NOT incrementing RTBD No match segments if it is disabled via servlet. Also, incrementing total requests count.  ![screen shot 2015-02-24 at 9 07 08 pm](https://github.corp.inmobi.com/github-enterprise-assets/0000/0575/0000/0053/42be9ec8-bcb3-11e4-970c-fb13343cf5dd.png)   NOTE: This pull request is dependent on thrift changes which is not yet merged. https://github.corp.inmobi.com/channel-adserve/cas-thrift/pull/84=yasir-imteyaz

+++ Server/src/main/java/com/inmobi/adserve/channels/server/servlet/ServletGetSegment.java
@@ -135,7 +135,7 @@ public class ServletGetSegment implements Servlet {
                     ResultSet<IXPackageEntity> resultSet =
                             CasConfigUtil.repositoryHelper.queryIXPackageRepository(Integer.parseInt(id.split("_")[0]),
                                     id.split("_")[1], Integer.parseInt(id.split("_")[2]), Integer.parseInt(id.split("_")[3]));
-                    List<Long> result = new ArrayList<>();
+                    List<Integer> result = new ArrayList<>();
                     for (IXPackageEntity packageEntity : resultSet) {
                         result.add(packageEntity.getId());
                     }
